fix l10

+++ modules/L10_system_emulator.sh
@@ -52,22 +52,26 @@ L10_system_emulator() {
           LIBNVRAM=$(get_nvram "$ARCH_END")
 
           create_emulation_filesystem "$R_PATH" "$ARCH_END"
-          identify_networking "$IMAGE_NAME" "$ARCH_END"
-          get_networking_details
+          if [[ "$FS_CREATED" -eq 1 ]]; then
+            identify_networking "$IMAGE_NAME" "$ARCH_END"
+            get_networking_details
 
-          if [[ "$KPANIC" -eq 0 && "${#IPS[@]}" -gt 0 ]]; then
-            setup_network
-            run_emulated_system
-            check_online_stat
-            if [[ "$SYS_ONLINE" -eq 1 ]]; then
-              print_output "[+] System emulation was successful."
-              print_output "[+] System should be available via IP $IP."
+            if [[ "$KPANIC" -eq 0 && "${#IPS[@]}" -gt 0 ]]; then
+              setup_network
+              run_emulated_system
+              check_online_stat
+              if [[ "$SYS_ONLINE" -eq 1 ]]; then
+                print_output "[+] System emulation was successful."
+                print_output "[+] System should be available via IP $IP."
+              else
+                reset_network
+              fi
+              create_emulation_archive
+              # if the emulation was successful, we stop here - no emulation of other detected rootfs
+              break
             else
-              reset_network
+              print_output "[!] No further emulation steps are performed"
             fi
-            create_emulation_archive
-            # if the emulation was successful, we stop here - no emulation of other detected rootfs
-            break
           else
             print_output "[!] No further emulation steps are performed"
           fi
@@ -98,7 +102,8 @@ create_emulation_filesystem() {
   ROOT_PATH="$1"
   ARCH_END="$2"
   export IMAGE_NAME
-  IMAGE_NAME="$(basename "$ROOT_PATH")_$ARCH_END"
+  FS_CREATED=1
+  IMAGE_NAME="$(basename "$ROOT_PATH")_$ARCH_END-$RANDOM"
   MNT_POINT="$LOG_PATH_MODULE/emulation_tmp_fs"
   if [[ -d "$MNT_POINT" ]]; then
     MNT_POINT="$MNT_POINT"-"$RANDOM"
@@ -112,7 +117,7 @@ create_emulation_filesystem() {
   print_output "[*] Size of filesystem for emulation - $IMAGE_SIZE.\\n"
   print_output "[*] Name of filesystem for emulation - $IMAGE_NAME.\\n"
   qemu-img create -f raw "$LOG_PATH_MODULE/$IMAGE_NAME" "$IMAGE_SIZE"
-  chmod a+rw "$LOG_PATH_MODULE/$IMAGE_NAME" | tee -a "$LOG_FILE"
+  chmod a+rw "$LOG_PATH_MODULE/$IMAGE_NAME"
 
   print_output "[*] Creating Partition Table"
   echo -e "o\nn\np\n1\n\n\nw" | /sbin/fdisk "$LOG_PATH_MODULE/$IMAGE_NAME"
@@ -126,42 +131,50 @@ create_emulation_filesystem() {
   mkfs.ext2 "${DEVICE}"
   sync
 
-  print_output "[*] Mounting QEMU Image Partition 1"
+  print_output "[*] Mounting QEMU Image Partition 1 to $MNT_POINT"
   mount "${DEVICE}" "$MNT_POINT"
+  if mount | grep -q "$MNT_POINT"; then
+    print_output "[*] Copy root filesystem to QEMU image"
+    #rm -rf "${MNT_POINT:?}/"*
+    cp -prf "$ROOT_PATH"/* "$MNT_POINT"/
 
-  print_output "[*] Copy root filesystem to QEMU image"
-  #rm -rf "${MNT_POINT:?}/"*
-  cp -prf "$ROOT_PATH"/* "$MNT_POINT"/
+    print_output "[*] Creating FIRMADYNE Directories"
+    mkdir -p "$MNT_POINT/firmadyne/libnvram/"
+    mkdir "$MNT_POINT/firmadyne/libnvram.override/"
 
-  print_output "[*] Creating FIRMADYNE Directories"
-  mkdir -p "$MNT_POINT/firmadyne/libnvram/"
-  mkdir "$MNT_POINT/firmadyne/libnvram.override/"
+    print_output "[*] Patching Filesystem (chroot)"
+    cp "$(which busybox)" "$MNT_POINT"
+    cp "$FIRMADYNE_DIR/scripts/fixImage.sh" "$MNT_POINT"
+    chroot "$MNT_POINT" /busybox ash /fixImage.sh
+    rm "$MNT_POINT/fixImage.sh"
+    rm "$MNT_POINT/busybox"
 
-  print_output "[*] Patching Filesystem (chroot)"
-  cp "$(which busybox)" "$MNT_POINT"
-  cp "$FIRMADYNE_DIR/scripts/fixImage.sh" "$MNT_POINT"
-  chroot "$MNT_POINT" /busybox ash /fixImage.sh
-  rm "$MNT_POINT/fixImage.sh"
-  rm "$MNT_POINT/busybox"
+    print_output "[*] Setting up FIRMADYNE"
+    cp "${CONSOLE}" "$MNT_POINT/firmadyne/console"
+    chmod a+x "$MNT_POINT/firmadyne/console"
+    mknod -m 666 "$MNT_POINT/firmadyne/ttyS1" c 4 65
 
-  print_output "[*] Setting up FIRMADYNE"
-  cp "${CONSOLE}" "$MNT_POINT/firmadyne/console"
-  chmod a+x "$MNT_POINT/firmadyne/console"
-  mknod -m 666 "$MNT_POINT/firmadyne/ttyS1" c 4 65
+    cp "${LIBNVRAM}" "$MNT_POINT/firmadyne/libnvram.so"
+    chmod a+x "$MNT_POINT/firmadyne/libnvram.so"
 
-  cp "${LIBNVRAM}" "$MNT_POINT/firmadyne/libnvram.so"
-  chmod a+x "$MNT_POINT/firmadyne/libnvram.so"
+    cp "$FIRMADYNE_DIR/scripts/preInit.sh" "$MNT_POINT/firmadyne/preInit.sh"
+    chmod a+x "$MNT_POINT/firmadyne/preInit.sh"
 
-  cp "$FIRMADYNE_DIR/scripts/preInit.sh" "$MNT_POINT/firmadyne/preInit.sh"
-  chmod a+x "$MNT_POINT/firmadyne/preInit.sh"
+    print_output "[*] Unmounting QEMU Image"
+    sync
+    umount "${DEVICE}"
 
-  print_output "[*] Unmounting QEMU Image"
-  sync
-  umount "${DEVICE}"
-
+  else
+    print_output "[!] Filesystem mount failed"
+    FS_CREATED=0
+  fi
   print_output "[*] Deleting device mapper"
-  kpartx -d "$LOG_PATH_MODULE/$IMAGE_NAME"
+  kpartx -v -d "$LOG_PATH_MODULE/$IMAGE_NAME"
   losetup -d "${DEVICE}" &>/dev/null
+  # just in case we check the output and remove our device:
+  if losetup | grep -q $(basename "$IMAGE_NAME"); then
+    losetup -d $(losetup | grep $(basename "$IMAGE_NAME") | awk '{print $1}')
+  fi
   dmsetup remove "$(basename "$DEVICE")" &>/dev/null
   rm -rf "${MNT_POINT:?}/"*
 }
